test(responses): fold the bridged streaming regressions into the mapped test file
